// assets/src/__generated__/OwnedComparisonSnapshotsQuery.graphql.ts
export type OwnedComparisonSnapshotsQuery$variables = {
  after?: string | null | undefined;
  first: number;
};
export type OwnedComparisonSnapshotsQuery$data = {
  readonly viewer: {
    readonly comparisonSnapshots: {
      readonly edges: ReadonlyArray<{
        readonly node: {
          readonly id: string;
          readonly sharePath: string;
          readonly title: string | null | undefined;
        };
      }>;
      readonly pageInfo: {
        readonly endCursor: string | null | undefined;
        readonly hasNextPage: boolean;
      };
    };
  } | null | undefined;
};
export type OwnedComparisonSnapshotsQuery = {
  response: OwnedComparisonSnapshotsQuery$data;
  variables: OwnedComparisonSnapshotsQuery$variables;
};